project components updated
