+++ src/components/project/banner/project-banner.module.css
@@ -1,0 +1,270 @@
+.root {
+  z-index: 10;
+  margin-left: auto;
+  margin-right: auto;
+  display: flex;
+  width: 83.333%;
+  flex-direction: column;
+  gap: 1rem;
+
+  @media (min-width: 768px) {
+    width: 66.666%;
+  }
+
+  @media (min-width: 1024px) {
+    width: 60%;
+  }
+
+  @media (min-width: 1536px) {
+    width: 55%;
+  }
+}
+
+.image-container {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  height: 208px;
+
+  @media (min-width: 640px) {
+    height: 256px;
+  }
+
+  @media (min-width: 768px) {
+    height: 384px;
+  }
+
+  @media (min-width: 1024px) {
+    height: 420px;
+  }
+
+  @media (min-width: 1536px) {
+    height: 470px;
+  }
+
+  @media (min-width: 1792px) {
+    height: 550px;
+  }
+}
+
+.image-wrapper {
+  position: relative;
+  height: 100%;
+  width: 100%;
+  border-radius: 1rem;
+  box-shadow: 0 10px 15px -3px rgb(0 0 0 / 0.3);
+  clip-path: inset(0 round 1rem);
+}
+
+.image-default {
+  position: absolute;
+  object-fit: cover;
+  object-position: 50% 70%;
+}
+
+.image-phone {
+  margin-top: 3.5rem;
+  transform: scale(1.5) rotate(-10deg);
+  object-fit: contain;
+
+  @media (min-width: 640px) {
+    margin-top: 5rem;
+    transform: scale(1.9) rotate(-10deg);
+  }
+
+  @media (min-width: 1024px) {
+    margin-left: -2.5rem;
+    margin-top: 6rem;
+    transform: scale(1.6) rotate(-10deg);
+  }
+
+  @media (min-width: 1280px) {
+    margin-top: 6rem;
+    margin-left: -4rem;
+  }
+
+  @media (min-width: 1792px) {
+    margin-top: 3.5rem;
+    margin-left: -7rem;
+  }
+}
+
+.image-pc {
+  margin-top: 3.5rem;
+  margin-left: -20%;
+  transform: scale(1.5) rotate(-5deg);
+  object-fit: contain;
+
+  @media (min-width: 460px) {
+    margin-left: -25%;
+    transform: scale(1.8) rotate(-5deg);
+  }
+
+  @media (min-width: 640px) {
+    margin-left: -30%;
+    margin-top: 5rem;
+    transform: scale(1.9) rotate(-5deg);
+  }
+
+  @media (min-width: 768px) {
+    margin-left: -25%;
+    margin-top: 6rem;
+    transform: scale(1.6) rotate(-5deg);
+  }
+
+  @media (min-width: 1024px) {
+    margin-top: 7rem;
+    margin-left: -25%;
+  }
+
+  @media (min-width: 1280px) {
+    margin-top: 7rem;
+    transform: scale(1.9) rotate(-5deg);
+  }
+
+  @media (min-width: 1536px) {
+    margin-top: 8rem;
+  }
+
+  @media (min-width: 1792px) {
+    margin-top: 9rem;
+  }
+}
+
+.image-phone2 {
+  margin-top: 2.5rem;
+  margin-left: 2.5rem;
+  transform: scale(1.7) rotate(-5deg);
+  object-fit: contain;
+
+  @media (min-width: 640px) {
+    margin-top: 1.75rem;
+    margin-left: 3.5rem;
+    transform: scale(1.9) rotate(-5deg);
+  }
+
+  @media (min-width: 768px) {
+    margin-left: -2rem;
+    margin-top: 6rem;
+    transform: scale(1.7) rotate(-5deg);
+  }
+
+  @media (min-width: 1024px) {
+    margin-top: 2.5rem;
+    margin-left: -3.5rem;
+    transform: scale(1.8) rotate(-5deg);
+  }
+
+  @media (min-width: 1792px) {
+    margin-left: -3rem;
+  }
+}
+
+.desktop-overlay {
+  position: absolute;
+  top: 1.25rem;
+  right: 2.5rem;
+  display: none;
+  height: 16.666%;
+  flex-direction: column;
+  gap: 0.5rem;
+
+  @media (min-width: 1024px) {
+    display: flex;
+
+    @media (prefers-color-scheme: dark) {
+      color: var(--color-foreground);
+    }
+  }
+
+  @media (min-width: 1024px) {
+    top: 1.5rem;
+    right: 3.5rem;
+    gap: 0.75rem;
+  }
+}
+
+.mobile-overlay {
+  display: flex;
+  width: 100%;
+  align-items: center;
+  justify-content: center;
+
+  @media (min-width: 1024px) {
+    display: none;
+  }
+}
+
+.title {
+  font-family: var(--font-heading);
+  font-weight: 500;
+  color: white;
+  font-size: 1.875rem;
+
+  @media (min-width: 1024px) {
+    font-size: 2.25rem;
+  }
+
+  @media (min-width: 1280px) {
+    font-size: 3rem;
+  }
+}
+
+.button-group {
+  margin-left: auto;
+  display: flex;
+  gap: 1rem;
+
+  @media (min-width: 1024px) {
+    gap: 1.5rem;
+  }
+}
+
+.button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 9999px;
+  background-color: white;
+  padding: 0.75rem;
+  transition-property: transform;
+  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  transition-duration: 300ms;
+
+  &:hover {
+    transform: scale(0.95);
+  }
+
+  @media (min-width: 640px) {
+    padding: 0.75rem;
+  }
+
+  @media (min-width: 1024px) {
+    padding: 0.75rem;
+  }
+
+  @media (min-width: 1792px) {
+    padding: 1rem;
+  }
+}
+
+.button-icon {
+  height: 0.75rem;
+  width: 0.75rem;
+  color: black;
+
+  @media (min-width: 640px) {
+    height: 1rem;
+    width: 1rem;
+  }
+
+  @media (min-width: 1024px) {
+    height: 1rem;
+    width: 1rem;
+  }
+
+  @media (min-width: 1792px) {
+    height: 1.25rem;
+    width: 1.25rem;
+  }
+}
