+++ src/components/project/description/project-description.module.css
@@ -1,0 +1,200 @@
+.root {
+  background-color: var(--color-whitebg);
+  width: 100%;
+  border-radius: 0 0 40px 40px;
+  padding-top: 2.5rem;
+  padding-bottom: 2.5rem;
+
+  @media (prefers-color-scheme: dark) {
+    background-color: #1a1a1a;
+  }
+
+  @media (min-width: 768px) {
+    padding-top: 3.5rem;
+    padding-bottom: 6rem;
+    border-radius: 0 0 50px 50px;
+  }
+
+  @media (min-width: 1024px) {
+    padding-bottom: 7rem;
+  }
+
+  @media (min-width: 1280px) {
+    padding-top: 5rem;
+    padding-bottom: 8rem;
+  }
+
+  @media (min-width: 1536px) {
+    border-radius: 0 0 70px 70px;
+  }
+}
+
+.content-wrapper {
+  margin-left: auto;
+  margin-right: auto;
+  display: flex;
+  width: 83.333%;
+  max-width: 1920px;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+.header {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.title {
+  font-family: var(--font-heading);
+  font-size: 1.5rem;
+  font-weight: 600;
+  line-height: 1;
+
+  @media (min-width: 640px) {
+    font-size: 1.875rem;
+  }
+
+  @media (min-width: 768px) {
+    font-size: 2.25rem;
+  }
+
+  @media (min-width: 1280px) {
+    font-size: 3rem;
+  }
+}
+
+.badges {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+
+  @media (min-width: 1280px) {
+    gap: 0.75rem;
+  }
+}
+
+.main-content {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+
+  @media (min-width: 1024px) {
+    flex-direction: row;
+    align-items: center;
+    gap: 3.5rem;
+  }
+
+  @media (min-width: 1536px) {
+    justify-content: space-between;
+    gap: 5rem;
+  }
+}
+
+.disclaimer {
+  font-family: var(--font-body);
+  margin-bottom: 1rem;
+  font-size: 0.875rem;
+  font-weight: 300;
+  line-height: 1.5rem;
+
+  @media (min-width: 768px) {
+    font-size: 1rem;
+  }
+}
+
+.disclaimer-label {
+  font-weight: 400;
+  font-style: italic;
+}
+
+.paragraph {
+  font-family: var(--font-body);
+  margin-bottom: 1rem;
+  font-size: 0.875rem;
+  font-weight: 300;
+  line-height: 1.5rem;
+
+  @media (min-width: 768px) {
+    font-size: 1rem;
+  }
+}
+
+.test-user-section {
+  margin-top: 1rem;
+}
+
+.test-user-heading {
+  font-family: var(--font-body);
+  font-size: 0.875rem;
+  font-weight: 400;
+  font-style: italic;
+  line-height: 1.5rem;
+
+  @media (min-width: 768px) {
+    font-size: 1rem;
+  }
+}
+
+.test-user-item {
+  font-family: var(--font-body);
+  font-size: 0.875rem;
+  font-weight: 300;
+  line-height: 1.5rem;
+
+  @media (min-width: 768px) {
+    font-size: 1rem;
+  }
+}
+
+.test-user-label {
+  font-weight: 400;
+  font-style: italic;
+}
+
+.divider {
+  background-color: var(--color-foreground);
+  height: 1px;
+  border: none;
+
+  @media (min-width: 1024px) {
+    display: none;
+  }
+}
+
+.tech-section {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  border-radius: 1rem;
+
+  @media (min-width: 1024px) {
+    align-self: start;
+    width: 100%;
+  }
+
+  @media (min-width: 1280px) {
+    width: 66.666%;
+  }
+
+  @media (min-width: 1792px) {
+    width: 50%;
+  }
+}
+
+.tech-title {
+  font-family: var(--font-heading);
+  display: none;
+  font-size: 1.5rem;
+  font-weight: 600;
+
+  @media (min-width: 1024px) {
+    display: block;
+  }
+}
+
+.tech-icons {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
