+++ src/components/project/index.ts
@@ -1,0 +1,3 @@
+export { ProjectCard } from "./card/project-card";
+export { ProjectBanner } from "./banner/project-banner";
+export { ProjectDescription } from "./description/project-description";
