+++ src/components/visuals/index.ts
@@ -1,0 +1,1 @@
+export { Dots } from "./dots/dots";
